count the number of correct people who get coats.
Six friends all throw their coats on the bed in the spare room when they show up for Teesha's birthday party. At the end of the party, they all enter the dark room and randomly grab a coat and put it on without knowing if it is actually their coats. What is the probability that exactly three of the friends got the right coat?